+++ TabletopTweaks-Base/Bugfixes/General/WeaponProperties.cs
@@ -41,13 +41,5 @@ namespace TabletopTweaks.Base.Bugfixes.General {
                 }
             }
         }
-        [HarmonyPatch(typeof(RuleCalculateWeaponStats), nameof(RuleCalculateWeaponStats.CriticalRange), MethodType.Getter)]
-        class RuleCalculateWeaponStats_CriticalRange_Patch {
-            static void Postfix(RuleCalculateWeaponStats __instance, ref int __result) {
-                if (TTTContext.Fixes.BaseFixes.IsDisabled("CriticalRangeIncreases")) { return; }
-                __result = ((21 - __instance.Weapon.Blueprint.CriticalRollEdge) * (__instance.DoubleCriticalEdge ? 2 : 1))
-                    + __instance.CriticalEdgeBonus;
-            }
-        }
     }
 }